add Paddle Trt config: with_interleaved
